refactor(reciptor-server): move all authentication-related classes to an authentication-package
