+++ docs/springBoot.md
@@ -9,3 +9,6 @@ $ (cd apps/reciptor-server && mvn flyway:migrate)
 
 If a migration is not recognized even though being correctly named according to [flyway-docs](https://flywaydb.org/documentation/concepts/migrations.html#sql-based-migrations)
 you should clear `/target`-folder and try again, since flyway will search in this folder for migrations.
+
+# weiter mit
+Next, create WebSecurityConfig and add EnableWebFluxSecurity and EnableReactiveMethodSecurty annotation, in this component you can configure all your security needs, like authentication manager, security context repository, which url is in permit (in this case /login), etc.
